Split renderer/main.tsx into per-tab modules + pdf-modal
- tabs/projects.tsx: ProjectsView, GroupBlock, Card, step helpers, drag MIME
- tabs/knowledge.tsx: KnowledgeView, KnowledgeNodeView, filter helper
- tabs/code.tsx: RepoRow, BranchActions, groupRepos, status helpers
- tabs/search.tsx: SearchResult
- pdf-modal.tsx: PdfModal

main.tsx shrinks from 1714 to 786 LOC (App + tab routing + modal plumbing).
No behaviour change; typecheck passes.
