+++ src/renderer/main.tsx
@@ -11,28 +11,35 @@ import {
 } from 'solid-js';
 import type {
   Deliverable,
-  KnowledgeNode,
   OpenWithSlotKey,
   OpenWithSlots,
   Project,
   RepoEntry,
   SearchHit,
   Step,
-  StepCounts,
-  StepMarker,
   TermSession,
   TerminalPrefs,
   Theme,
   TreeEvent,
   Worktree,
 } from '@shared/types';
-import { KNOWN_STATUSES, STEP_MARKERS } from '@shared/types';
 import { CodeRunRows } from './code-runs';
 import { NoteModal, type ModalState } from './note-modal';
 import { ProjectPreview } from './project-preview';
 import { resetMermaidTheme } from './markdown';
 import { TerminalPane, type TerminalPaneHandle } from './terminal-pane';
 import { buildSlugIndex } from './wikilinks';
+import { PdfModal } from './pdf-modal';
+import {
+  applyStatus,
+  applyStepMarker,
+  groupByStatus,
+  nextMarker,
+  ProjectsView,
+} from './tabs/projects';
+import { KnowledgeView } from './tabs/knowledge';
+import { groupRepos, RepoRow } from './tabs/code';
+import { SearchResult } from './tabs/search';
 import './styles.css';
 import './note-modal.css';
 import './project-preview.css';
@@ -46,20 +53,6 @@ const THEME_LABEL: Record<Theme, string> = {
   dark: '☾ dark',
 };
 
-const MARKER_GLYPH: Record<StepMarker, string> = {
-  ' ': '☐',
-  '~': '◐',
-  x: '☑',
-  '-': '✕',
-};
-
-const MARKER_LABEL: Record<StepMarker, string> = {
-  ' ': 'todo',
-  '~': 'doing',
-  x: 'done',
-  '-': 'dropped',
-};
-
 function applyTheme(theme: Theme): void {
   const root = document.documentElement;
   if (theme === 'system') {
@@ -67,203 +60,6 @@ function applyTheme(theme: Theme): void {
   } else {
     root.setAttribute('data-theme', theme);
   }
-}
-
-function hasSteps(c: StepCounts): boolean {
-  return c.todo + c.doing + c.done + c.dropped > 0;
-}
-
-/** First step whose marker is not 'x' (done). Returns undefined if every step
- * is done — the card body collapses in that case. */
-function nextOpenStep(item: Project): Step | undefined {
-  return item.steps.find((s) => s.marker !== 'x');
-}
-
-function StepProgress(props: { counts: StepCounts }) {
-  const total = (): number =>
-    props.counts.todo + props.counts.doing + props.counts.done + props.counts.dropped;
-  const ratio = (): number => {
-    const t = total();
-    return t === 0 ? 0 : Math.min(1, props.counts.done / t);
-  };
-  const title = (): string =>
-    `${props.counts.todo} todo, ${props.counts.doing} doing, ${props.counts.done} done, ${props.counts.dropped} dropped`;
-  return (
-    <span class="step-progress-inner" title={title()}>
-      <span class="progress-track">
-        <span class="progress-fill" style={{ width: `${ratio() * 100}%` }} />
-      </span>
-      <span class="progress-text">
-        {props.counts.done}/{total()}
-      </span>
-    </span>
-  );
-}
-
-const KIND_ICON: Record<string, () => any> = {
-  project: () => (
-    <svg
-      viewBox="0 0 16 16"
-      fill="none"
-      stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      aria-hidden="true"
-    >
-      <path d="M8 1.5L14.5 8 8 14.5 1.5 8z" />
-    </svg>
-  ),
-  incident: () => (
-    <svg
-      viewBox="0 0 16 16"
-      fill="none"
-      stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      aria-hidden="true"
-    >
-      <path d="M8 2L14.5 13.5h-13z" />
-      <path d="M8 6.5v3" />
-      <circle cx="8" cy="11.5" r="0.4" fill="currentColor" stroke="none" />
-    </svg>
-  ),
-  document: () => (
-    <svg
-      viewBox="0 0 16 16"
-      fill="none"
-      stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      aria-hidden="true"
-    >
-      <path d="M3.5 1.5h6L13 5v9.5H3.5z" />
-      <path d="M9.5 1.5V5H13" />
-      <path d="M5.5 8h5M5.5 10.5h5M5.5 5.5h2" />
-    </svg>
-  ),
-};
-
-function KindIcon(props: { kind: string }) {
-  const Icon = KIND_ICON[props.kind];
-  if (!Icon) return null;
-  return <Icon />;
-}
-
-function AppsIcon() {
-  return (
-    <svg
-      viewBox="0 0 16 16"
-      fill="none"
-      stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      aria-hidden="true"
-    >
-      <path d="M2 4.5l6-3 6 3-6 3z" />
-      <path d="M2 8l6 3 6-3" />
-      <path d="M2 11.5l6 3 6-3" />
-    </svg>
-  );
-}
-
-function DownloadIcon() {
-  return (
-    <svg
-      viewBox="0 0 16 16"
-      fill="none"
-      stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      aria-hidden="true"
-    >
-      <path d="M8 2v8" />
-      <path d="M4.5 7L8 10.5 11.5 7" />
-      <path d="M2.5 13.5h11" />
-    </svg>
-  );
-}
-
-function WarnIcon() {
-  return (
-    <svg
-      viewBox="0 0 16 16"
-      fill="none"
-      stroke="currentColor"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-      aria-hidden="true"
-    >
-      <circle cx="8" cy="8" r="6.5" />
-      <path d="M8 5v3.5" />
-      <circle cx="8" cy="11" r="0.4" fill="currentColor" stroke="none" />
-    </svg>
-  );
-}
-
-function nextMarker(current: StepMarker): StepMarker {
-  const idx = STEP_MARKERS.indexOf(current);
-  return STEP_MARKERS[(idx + 1) % STEP_MARKERS.length];
-}
-
-function applyStepMarker(
-  items: Project[],
-  path: string,
-  lineIndex: number,
-  marker: StepMarker,
-): Project[] {
-  return items.map((p) => {
-    if (p.path !== path) return p;
-    const steps = p.steps.map((s) => (s.lineIndex === lineIndex ? { ...s, marker } : s));
-    return { ...p, steps, stepCounts: countSteps(steps) };
-  });
-}
-
-function countSteps(steps: readonly Step[]): StepCounts {
-  const c: StepCounts = { todo: 0, doing: 0, done: 0, dropped: 0 };
-  for (const s of steps) {
-    if (s.marker === ' ') c.todo++;
-    else if (s.marker === '~') c.doing++;
-    else if (s.marker === 'x') c.done++;
-    else if (s.marker === '-') c.dropped++;
-  }
-  return c;
-}
-
-function applyStatus(items: Project[], path: string, status: string): Project[] {
-  return items.map((p) => (p.path === path ? { ...p, status } : p));
-}
-
-const DRAG_MIME = 'application/x-condash-project-path';
-
-type Group = { status: string; items: Project[] };
-
-const UNKNOWN = '?';
-
-/** Order of stacked sections on the Projects tab. `backlog` and `done`
- * render collapsed-by-default — heavy buckets the user usually skips past. */
-const PROJECT_SECTION_ORDER = ['now', 'review', 'soon', 'later', 'backlog', 'done'] as const;
-const COLLAPSED_BY_DEFAULT = new Set<string>(['backlog', 'done']);
-
-function groupByStatus(items: Project[]): Map<string, Project[]> {
-  const buckets = new Map<string, Project[]>();
-  for (const status of KNOWN_STATUSES) buckets.set(status, []);
-  buckets.set(UNKNOWN, []);
-
-  for (const item of items) {
-    const key = (KNOWN_STATUSES as readonly string[]).includes(item.status) ? item.status : UNKNOWN;
-    buckets.get(key)!.push(item);
-  }
-  return buckets;
-}
-
-function projectsTabGroups(buckets: Map<string, Project[]>): Group[] {
-  const out: Group[] = [];
-  for (const status of PROJECT_SECTION_ORDER) {
-    out.push({ status, items: buckets.get(status) ?? [] });
-  }
-  const unknown = buckets.get(UNKNOWN) ?? [];
-  if (unknown.length > 0) out.push({ status: UNKNOWN, items: unknown });
-  return out;
 }
 
 interface Shortcut {
@@ -981,730 +777,6 @@ function App() {
           {toast()}
         </div>
       </Show>
-    </div>
-  );
-}
-
-function projectMatches(item: Project, needle: string): boolean {
-  if (!needle) return true;
-  return (
-    item.title.toLowerCase().includes(needle) ||
-    item.slug.toLowerCase().includes(needle) ||
-    (item.summary?.toLowerCase().includes(needle) ?? false) ||
-    (item.apps?.toLowerCase().includes(needle) ?? false)
-  );
-}
-
-function ProjectsView(props: {
-  buckets: Map<string, Project[]>;
-  onOpen: (project: Project) => void;
-  onToggleStep: (project: Project, step: Step) => void;
-  onDropProject: (path: string, newStatus: string) => void;
-}) {
-  const [filter, setFilter] = createSignal('');
-  const trimmedQuery = createMemo(() => filter().trim().toLowerCase());
-  const filteredGroups = createMemo<Group[]>(() => {
-    const q = trimmedQuery();
-    const buckets = props.buckets;
-    if (!q) return projectsTabGroups(buckets);
-    const filtered = new Map<string, Project[]>();
-    for (const [status, items] of buckets) {
-      filtered.set(
-        status,
-        items.filter((it) => projectMatches(it, q)),
-      );
-    }
-    return projectsTabGroups(filtered);
-  });
-  return (
-    <div class="projects-stack">
-      <div class="projects-filter">
-        <input
-          class="projects-filter-input"
-          type="search"
-          placeholder="Filter projects (title, slug, app, summary)…"
-          value={filter()}
-          onInput={(e) => setFilter(e.currentTarget.value)}
-        />
-      </div>
-      <For each={filteredGroups()}>
-        {(group) => (
-          <GroupBlock
-            group={group}
-            collapsedByDefault={COLLAPSED_BY_DEFAULT.has(group.status)}
-            forceOpen={trimmedQuery().length > 0 && group.items.length > 0}
-            onOpen={props.onOpen}
-            onToggleStep={props.onToggleStep}
-            onDropProject={props.onDropProject}
-          />
-        )}
-      </For>
-    </div>
-  );
-}
-
-function GroupBlock(props: {
-  group: Group;
-  /** When true, the section starts collapsed and shows an expand affordance. */
-  collapsedByDefault?: boolean;
-  /** Override collapsed state — e.g. when a search filter is active and the
-   * group has matches, force it open so results aren't hidden. */
-  forceOpen?: boolean;
-  onOpen: (project: Project) => void;
-  onToggleStep: (project: Project, step: Step) => void;
-  onDropProject: (path: string, newStatus: string) => void;
-}) {
-  const [over, setOver] = createSignal(false);
-  const [userExpanded, setUserExpanded] = createSignal<boolean | null>(null);
-  const isOpen = (): boolean => {
-    if (props.forceOpen) return true;
-    const ux = userExpanded();
-    if (ux !== null) return ux;
-    return !props.collapsedByDefault;
-  };
-
-  const isAcceptable = (event: DragEvent): boolean => {
-    const types = event.dataTransfer?.types;
-    return types ? Array.from(types).includes(DRAG_MIME) : false;
-  };
-
-  const handleDragEnter = (e: DragEvent) => {
-    if (!isAcceptable(e)) return;
-    e.preventDefault();
-    setOver(true);
-  };
-
-  const handleDragOver = (e: DragEvent) => {
-    if (!isAcceptable(e)) return;
-    e.preventDefault();
-    if (e.dataTransfer) e.dataTransfer.dropEffect = 'move';
-  };
-
-  const handleDragLeave = (e: DragEvent) => {
-    if (e.currentTarget === e.target) setOver(false);
-  };
-
-  const handleDrop = (e: DragEvent) => {
-    if (!isAcceptable(e)) return;
-    e.preventDefault();
-    setOver(false);
-    const path = e.dataTransfer?.getData(DRAG_MIME);
-    if (path) props.onDropProject(path, props.group.status);
-  };
-
-  return (
-    <section
-      class="group-block"
-      classList={{ 'drag-over': over(), collapsed: !isOpen() }}
-      data-status={props.group.status}
-      onDragEnter={handleDragEnter}
-      onDragOver={handleDragOver}
-      onDragLeave={handleDragLeave}
-      onDrop={handleDrop}
-    >
-      <header
-        class="group-header"
-        onClick={() => setUserExpanded(!isOpen())}
-        title={isOpen() ? 'Collapse section' : 'Expand section'}
-      >
-        <span class="caret" aria-hidden="true">
-          {isOpen() ? '▾' : '▸'}
-        </span>
-        <span class="dot" aria-hidden="true" />
-        <span class="name">{props.group.status}</span>
-        <span class="count">{props.group.items.length}</span>
-        <span class="rule" aria-hidden="true" />
-      </header>
-      <Show when={isOpen()}>
-        <div class="group-body">
-          <For each={props.group.items}>
-            {(item) => <Card item={item} onOpen={props.onOpen} onToggleStep={props.onToggleStep} />}
-          </For>
-        </div>
-      </Show>
-    </section>
-  );
-}
-
-function Card(props: {
-  item: Project;
-  onOpen: (project: Project) => void;
-  onToggleStep: (project: Project, step: Step) => void;
-  draggable?: boolean;
-}) {
-  const [expanded, setExpanded] = createSignal(false);
-
-  const handleHeaderClick = (event: MouseEvent) => {
-    if ((event.target as HTMLElement).closest('.step-toggle, .expander')) return;
-    props.onOpen(props.item);
-  };
-
-  const handleDragStart = (event: DragEvent) => {
-    if (!event.dataTransfer) return;
-    event.dataTransfer.setData(DRAG_MIME, props.item.path);
-    event.dataTransfer.effectAllowed = 'move';
-  };
-
-  const isDraggable = (): boolean => props.draggable !== false;
-
-  return (
-    <article
-      class="row"
-      title={props.item.path}
-      data-status-card={props.item.status}
-      draggable={isDraggable()}
-      onDragStart={isDraggable() ? handleDragStart : undefined}
-    >
-      <div class="row-head" onClick={handleHeaderClick}>
-        <h3 class="title">
-          <Show when={props.item.kind !== 'unknown'}>
-            <span class="title-kind" data-kind={props.item.kind} title={props.item.kind}>
-              <KindIcon kind={props.item.kind} />
-            </span>
-          </Show>
-          <span class="title-text">{props.item.title}</span>
-        </h3>
-        <Show when={nextOpenStep(props.item)} keyed>
-          {(step) => (
-            <p class="summary next-step" data-marker={markerClass(step.marker)}>
-              <span class="next-step-marker" aria-hidden="true">
-                {step.marker === '~' ? '◐' : step.marker === '-' ? '⨯' : '○'}
-              </span>
-              {step.text}
-            </p>
-          )}
-        </Show>
-        <div class="meta">
-          <Show when={props.item.apps}>
-            <span class="meta-icon apps" title={props.item.apps}>
-              <AppsIcon />
-              {props.item.apps}
-            </span>
-          </Show>
-          <Show when={props.item.deliverableCount > 0}>
-            <span class="meta-icon" title="deliverables">
-              <DownloadIcon />
-              {props.item.deliverableCount}
-            </span>
-          </Show>
-          <Show when={!(KNOWN_STATUSES as readonly string[]).includes(props.item.status)}>
-            <span class="meta-icon warn" title={`Unknown status: ${props.item.status}`}>
-              <WarnIcon />
-              {props.item.status}
-            </span>
-          </Show>
-          <span class="slug">{props.item.slug}</span>
-          <Show when={hasSteps(props.item.stepCounts)}>
-            <button
-              class="meta-icon expander step-bar-right"
-              onClick={(e) => {
-                e.stopPropagation();
-                setExpanded((v) => !v);
-              }}
-              title={`${props.item.steps.length} steps · click to ${expanded() ? 'collapse' : 'expand'}`}
-            >
-              <StepProgress counts={props.item.stepCounts} />
-              <span class="expander-arrow">{expanded() ? '▾' : '▸'}</span>
-            </button>
-          </Show>
-        </div>
-      </div>
-      <Show when={expanded() && props.item.steps.length > 0}>
-        <ul class="steps-list">
-          <For each={props.item.steps}>
-            {(step) => (
-              <li class={`step step-marker-${markerClass(step.marker)}`}>
-                <button
-                  class="step-toggle"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    props.onToggleStep(props.item, step);
-                  }}
-                  title={`${MARKER_LABEL[step.marker]} → ${MARKER_LABEL[nextMarker(step.marker)]}`}
-                >
-                  {MARKER_GLYPH[step.marker]}
-                </button>
-                <span class="step-text">{step.text}</span>
-              </li>
-            )}
-          </For>
-        </ul>
-      </Show>
-    </article>
-  );
-}
-
-/** Prune `node` to only the subtree where some descendant title or path
- * matches `needle`. Returns null when nothing in the subtree matches. */
-function filterKnowledgeTree(node: KnowledgeNode, needle: string): KnowledgeNode | null {
-  if (!needle) return node;
-  const titleHit = node.title.toLowerCase().includes(needle);
-  const pathHit = node.path.toLowerCase().includes(needle);
-  if (node.kind === 'file') {
-    return titleHit || pathHit ? node : null;
-  }
-  const children = (node.children ?? [])
-    .map((c) => filterKnowledgeTree(c, needle))
-    .filter((c): c is KnowledgeNode => c !== null);
-  if (children.length === 0 && !titleHit && !pathHit) return null;
-  return { ...node, children };
-}
-
-function KnowledgeView(props: { root: KnowledgeNode; onOpen: (path: string) => void }) {
-  const [filter, setFilter] = createSignal('');
-  const trimmed = createMemo(() => filter().trim().toLowerCase());
-  const filteredRoot = createMemo<KnowledgeNode | null>(() =>
-    filterKnowledgeTree(props.root, trimmed()),
-  );
-  return (
-    <div class="knowledge-pane">
-      <div class="projects-filter">
-        <input
-          class="projects-filter-input"
-          type="search"
-          placeholder="Filter knowledge (title, path)…"
-          value={filter()}
-          onInput={(e) => setFilter(e.currentTarget.value)}
-        />
-      </div>
-      <Show when={filteredRoot()} fallback={<div class="empty">No knowledge entries match.</div>}>
-        <ul class="knowledge-tree knowledge-tree-root">
-          <KnowledgeNodeView
-            node={filteredRoot()!}
-            depth={0}
-            onOpen={props.onOpen}
-            initiallyExpanded
-            forceExpand={trimmed().length > 0}
-          />
-        </ul>
-      </Show>
-    </div>
-  );
-}
-
-function KnowledgeNodeView(props: {
-  node: KnowledgeNode;
-  depth: number;
-  onOpen: (path: string) => void;
-  initiallyExpanded?: boolean;
-  /** When true, ignore the local toggle and expand — used so a search filter
-   * surfaces matches no matter how the user previously collapsed branches. */
-  forceExpand?: boolean;
-}) {
-  const [expanded, setExpanded] = createSignal(props.initiallyExpanded ?? props.depth === 0);
-  const isExpanded = (): boolean => props.forceExpand || expanded();
-
-  return (
-    <li class="knowledge-node" data-kind={props.node.kind}>
-      <Show
-        when={props.node.kind === 'directory'}
-        fallback={
-          <button
-            class="knowledge-leaf"
-            onClick={() => props.onOpen(props.node.path)}
-            title={props.node.path}
-          >
-            <span class="knowledge-icon">📄</span>
-            <span class="knowledge-title">{props.node.title}</span>
-          </button>
-        }
-      >
-        <button class="knowledge-dir" onClick={() => setExpanded((v) => !v)}>
-          <span class="knowledge-icon">{isExpanded() ? '▾' : '▸'}</span>
-          <span class="knowledge-title">{props.node.title}</span>
-          <Show when={props.node.children}>
-            <span class="knowledge-count">{props.node.children!.length}</span>
-          </Show>
-        </button>
-        <Show when={isExpanded() && props.node.children}>
-          <ul class="knowledge-tree">
-            <For each={props.node.children}>
-              {(child) => (
-                <KnowledgeNodeView
-                  node={child}
-                  depth={props.depth + 1}
-                  onOpen={props.onOpen}
-                  forceExpand={props.forceExpand}
-                />
-              )}
-            </For>
-          </ul>
-        </Show>
-      </Show>
-    </li>
-  );
-}
-
-type RepoGroup = { id: string; label: string; entries: RepoEntry[] };
-
-function groupRepos(repos: readonly RepoEntry[]): RepoGroup[] {
-  const childrenByParent = new Map<string, RepoEntry[]>();
-  for (const r of repos) {
-    if (!r.parent) continue;
-    const arr = childrenByParent.get(r.parent) ?? [];
-    arr.push(r);
-    childrenByParent.set(r.parent, arr);
-  }
-  const primary: RepoEntry[] = [];
-  const secondary: RepoEntry[] = [];
-  const submoduleParents: { parent: RepoEntry; children: RepoEntry[] }[] = [];
-  for (const r of repos) {
-    if (r.parent) continue;
-    const kids = childrenByParent.get(r.name);
-    if (kids && kids.length > 0) {
-      submoduleParents.push({ parent: r, children: kids });
-    } else if (r.kind === 'primary') {
-      primary.push(r);
-    } else {
-      secondary.push(r);
-    }
-  }
-  const groups: RepoGroup[] = [];
-  if (primary.length > 0) groups.push({ id: 'primary', label: 'PRIMARY', entries: primary });
-  for (const { parent, children } of submoduleParents) {
-    groups.push({
-      id: parent.name,
-      label: parent.name.toUpperCase(),
-      entries: [parent, ...children],
-    });
-  }
-  if (secondary.length > 0)
-    groups.push({ id: 'secondary', label: 'SECONDARY', entries: secondary });
-  return groups;
-}
-
-type RepoStatus = 'missing' | 'unknown' | 'clean' | 'dirty';
-
-/** Synthesise the primary checkout as a Worktree-shaped row when the data
- * layer didn't return any worktrees (e.g. repo missing or git failed). The
- * branch is unknown so we leave it null. */
-function ensureWorktrees(repo: RepoEntry): Worktree[] {
-  if (repo.worktrees && repo.worktrees.length > 0) return repo.worktrees;
-  return [
-    {
-      path: repo.path,
-      branch: null,
-      primary: true,
-      dirty: repo.dirty,
-    },
-  ];
-}
-
-/** Sort: primary checkout first, then worktrees alphabetically. */
-function orderedWorktrees(repo: RepoEntry): Worktree[] {
-  const list = ensureWorktrees(repo).slice();
-  list.sort((a, b) => {
-    if (a.primary !== b.primary) return a.primary ? -1 : 1;
-    return (a.branch ?? '').localeCompare(b.branch ?? '');
-  });
-  return list;
-}
-
-/** Per-card status pill text — collapses worktree dirtiness into one of:
- * "CLEAN", "1 BRANCH DIRTY", "N BRANCHES DIRTY", "MISSING", "?". */
-function cardStatusLabel(repo: RepoEntry): string {
-  if (repo.missing) return 'MISSING';
-  const wts = ensureWorktrees(repo);
-  const dirtyCount = wts.filter((w) => (w.dirty ?? 0) > 0).length;
-  if (dirtyCount === 0) {
-    if (wts.every((w) => w.dirty === null || w.dirty === undefined)) return '?';
-    return 'CLEAN';
-  }
-  return dirtyCount === 1 ? '1 BRANCH DIRTY' : `${dirtyCount} BRANCHES DIRTY`;
-}
-
-function cardStatus(repo: RepoEntry): RepoStatus {
-  if (repo.missing) return 'missing';
-  const wts = ensureWorktrees(repo);
-  if (wts.every((w) => w.dirty === null || w.dirty === undefined)) return 'unknown';
-  return wts.some((w) => (w.dirty ?? 0) > 0) ? 'dirty' : 'clean';
-}
-
-function RepoRow(props: {
-  repo: RepoEntry;
-  slots: OpenWithSlots;
-  /** True when at least one terminal session is currently running for this repo. */
-  live?: boolean;
-  onOpen: (path: string) => void;
-  onLaunch: (slot: OpenWithSlotKey, path: string) => void;
-  onForceStop: (repo: RepoEntry) => void;
-  onStop: (repo: RepoEntry) => void;
-  onRun: (repo: RepoEntry, worktree?: Worktree) => void;
-}) {
-  const status = (): RepoStatus => cardStatus(props.repo);
-  const displayName = (): string => {
-    if (props.repo.parent && props.repo.name.startsWith(`${props.repo.parent}/`)) {
-      return props.repo.name.slice(props.repo.parent.length + 1);
-    }
-    return props.repo.name;
-  };
-
-  const branchStatus = (wt: Worktree): RepoStatus => {
-    if (props.repo.missing) return 'missing';
-    if (wt.dirty == null) return 'unknown';
-    return wt.dirty === 0 ? 'clean' : 'dirty';
-  };
-
-  const hasRun = (): boolean => !props.repo.missing;
-
-  return (
-    <article
-      class="repo-row"
-      classList={{
-        missing: props.repo.missing,
-        submodule: !!props.repo.parent,
-      }}
-      data-status={status()}
-    >
-      <header class="repo-head">
-        <span class="repo-name">{displayName()}</span>
-        <span class="repo-status-badge" data-status={status()}>
-          {cardStatusLabel(props.repo)}
-        </span>
-        <Show when={props.live}>
-          <span class="repo-live-badge" title="A terminal session is running for this repo">
-            LIVE
-          </span>
-        </Show>
-        <Show when={props.live && props.repo.hasForceStop}>
-          <button
-            class="repo-action stop repo-stop-button"
-            onClick={() => props.onStop(props.repo)}
-            title={`Stop the running session for ${props.repo.name} (runs force_stop)`}
-            aria-label={`Stop ${props.repo.name}`}
-          >
-            ⏹
-          </button>
-        </Show>
-        <span class="spacer" />
-        <span class="repo-kind-tag" title={`Configured under repositories.${props.repo.kind}`}>
-          {props.repo.parent ? 'SUB' : 'REPO'}
-        </span>
-      </header>
-      <ul class="branches">
-        <For each={orderedWorktrees(props.repo)}>
-          {(wt) => (
-            <li class="branch-row" data-status={branchStatus(wt)}>
-              <span class="branch-dot" aria-hidden="true" />
-              <span class="branch-name">{wt.branch ?? '(detached)'}</span>
-              <span class="branch-role">{wt.primary ? 'CHECKOUT' : 'WORKTREE'}</span>
-              <Show when={(wt.dirty ?? 0) > 0}>
-                <span class="branch-dirty">{wt.dirty} dirty</span>
-              </Show>
-              <Show when={wt.primary && props.live}>
-                <span class="branch-live-dot" title="Running" aria-label="Running" />
-              </Show>
-              <span class="spacer" />
-              <BranchActions
-                repo={props.repo}
-                worktree={wt}
-                slots={props.slots}
-                hasRun={hasRun()}
-                onOpen={props.onOpen}
-                onLaunch={props.onLaunch}
-                onForceStop={props.onForceStop}
-                onRun={props.onRun}
-              />
-            </li>
-          )}
-        </For>
-      </ul>
-      <span class="repo-path" title={props.repo.path}>
-        {props.repo.path}
-      </span>
-    </article>
-  );
-}
-
-/** Per-branch action group: a primary Run + Open pair always visible, plus
- * the editor / terminal launchers and force_stop tucked behind a ▼ menu so
- * the row stays compact at narrower card widths. */
-function BranchActions(props: {
-  repo: RepoEntry;
-  worktree: Worktree;
-  slots: OpenWithSlots;
-  hasRun: boolean;
-  onOpen: (path: string) => void;
-  onLaunch: (slot: OpenWithSlotKey, path: string) => void;
-  onForceStop: (repo: RepoEntry) => void;
-  onRun: (repo: RepoEntry, worktree?: Worktree) => void;
-}) {
-  const [menuOpen, setMenuOpen] = createSignal(false);
-  let menuRoot: HTMLDivElement | undefined;
-
-  const onDocClick = (e: MouseEvent) => {
-    if (!menuOpen()) return;
-    if (menuRoot && !menuRoot.contains(e.target as Node)) setMenuOpen(false);
-  };
-  onMount(() => document.addEventListener('click', onDocClick, true));
-  onCleanup(() => document.removeEventListener('click', onDocClick, true));
-
-  const showForceStop = (): boolean => props.worktree.primary && !!props.repo.hasForceStop;
-  const launcherEntries = (): OpenWithSlotKey[] =>
-    LAUNCHER_SLOTS.filter((slot) => !!props.slots[slot]);
-  const hasOverflow = (): boolean => showForceStop() || launcherEntries().length > 0;
-
-  return (
-    <div class="branch-actions" ref={(el) => (menuRoot = el)}>
-      <Show when={props.hasRun}>
-        <button
-          class="repo-action run"
-          onClick={() => props.onRun(props.repo, props.worktree)}
-          disabled={props.repo.missing}
-          title={
-            props.worktree.primary
-              ? 'Run configured run: command'
-              : `Run configured run: command in ${props.worktree.branch ?? '(detached)'}`
-          }
-        >
-          ▶
-        </button>
-      </Show>
-      <button
-        class="repo-action icon"
-        onClick={() => props.onOpen(props.worktree.path)}
-        disabled={props.repo.missing}
-        title="Open in OS file manager"
-      >
-        📁
-      </button>
-      <Show when={hasOverflow()}>
-        <button
-          class="repo-action icon"
-          onClick={(e) => {
-            e.stopPropagation();
-            setMenuOpen((v) => !v);
-          }}
-          aria-haspopup="menu"
-          aria-expanded={menuOpen()}
-          title="More actions"
-        >
-          ▾
-        </button>
-      </Show>
-      <Show when={menuOpen() && hasOverflow()}>
-        <div class="branch-action-menu" role="menu">
-          <For each={launcherEntries()}>
-            {(slot) => (
-              <button
-                class="branch-action-menu-item"
-                role="menuitem"
-                onClick={() => {
-                  setMenuOpen(false);
-                  props.onLaunch(slot, props.worktree.path);
-                }}
-              >
-                <span class="glyph">{LAUNCHER_GLYPH[slot]}</span>
-                <span>{props.slots[slot]!.label}</span>
-              </button>
-            )}
-          </For>
-          <Show when={showForceStop()}>
-            <button
-              class="branch-action-menu-item warn"
-              role="menuitem"
-              onClick={() => {
-                setMenuOpen(false);
-                props.onForceStop(props.repo);
-              }}
-            >
-              <span class="glyph">⏹</span>
-              <span>Force-stop</span>
-            </button>
-          </Show>
-        </div>
-      </Show>
-    </div>
-  );
-}
-
-const LAUNCHER_SLOTS: readonly OpenWithSlotKey[] = ['main_ide', 'secondary_ide', 'terminal'];
-const LAUNCHER_GLYPH: Record<OpenWithSlotKey, string> = {
-  main_ide: '⌘',
-  secondary_ide: '⌥',
-  terminal: '▶',
-};
-
-function SearchResult(props: { hit: SearchHit; onOpen: (path: string) => void }) {
-  return (
-    <li class="search-result">
-      <button class="search-row" onClick={() => props.onOpen(props.hit.path)}>
-        <div class="search-head">
-          <span class="search-title">{props.hit.title}</span>
-          <span class="badge">{props.hit.source}</span>
-          <span class="search-count">{props.hit.matchCount}</span>
-        </div>
-        <span class="search-path">{props.hit.path}</span>
-        <ul class="search-snippets">
-          <For each={props.hit.snippets}>{(s) => <li>{s}</li>}</For>
-        </ul>
-      </button>
-    </li>
-  );
-}
-
-function markerClass(m: StepMarker): string {
-  if (m === ' ') return 'todo';
-  if (m === '~') return 'doing';
-  if (m === 'x') return 'done';
-  return 'dropped';
-}
-
-function PdfModal(props: {
-  path: string;
-  onClose: () => void;
-  onOpenInOs: (path: string) => void;
-}) {
-  const handleKey = (e: KeyboardEvent) => {
-    if (e.key === 'Escape') {
-      e.preventDefault();
-      props.onClose();
-    }
-  };
-
-  onMount(() => {
-    document.addEventListener('keydown', handleKey, true);
-  });
-  onCleanup(() => {
-    document.removeEventListener('keydown', handleKey, true);
-  });
-
-  const fileUrl = (): string => {
-    const encoded = props.path
-      .split('/')
-      .map((seg) => encodeURIComponent(seg))
-      .join('/');
-    return `file://${encoded}`;
-  };
-
-  const fileName = (): string => props.path.split('/').pop() ?? props.path;
-
-  return (
-    <div class="modal-backdrop" onClick={props.onClose}>
-      <div
-        class="modal pdf-modal"
-        role="dialog"
-        aria-modal="true"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <header class="modal-head">
-          <span class="modal-title">{fileName()}</span>
-          <span class="modal-path">{props.path}</span>
-          <button
-            class="modal-button"
-            onClick={() => props.onOpenInOs(props.path)}
-            title="Open in OS default viewer"
-          >
-            ↗
-          </button>
-          <button class="modal-button" onClick={props.onClose} title="Close (Esc)">
-            ×
-          </button>
-        </header>
-        <div class="pdf-body">
-          <webview src={fileUrl()} partition="persist:pdf" class="pdf-webview" />
-        </div>
-      </div>
     </div>
   );
 }
