+++ src/renderer/pdf-modal.tsx
@@ -1,0 +1,60 @@
+import { onCleanup, onMount } from 'solid-js';
+
+export function PdfModal(props: {
+  path: string;
+  onClose: () => void;
+  onOpenInOs: (path: string) => void;
+}) {
+  const handleKey = (e: KeyboardEvent) => {
+    if (e.key === 'Escape') {
+      e.preventDefault();
+      props.onClose();
+    }
+  };
+
+  onMount(() => {
+    document.addEventListener('keydown', handleKey, true);
+  });
+  onCleanup(() => {
+    document.removeEventListener('keydown', handleKey, true);
+  });
+
+  const fileUrl = (): string => {
+    const encoded = props.path
+      .split('/')
+      .map((seg) => encodeURIComponent(seg))
+      .join('/');
+    return `file://${encoded}`;
+  };
+
+  const fileName = (): string => props.path.split('/').pop() ?? props.path;
+
+  return (
+    <div class="modal-backdrop" onClick={props.onClose}>
+      <div
+        class="modal pdf-modal"
+        role="dialog"
+        aria-modal="true"
+        onClick={(e) => e.stopPropagation()}
+      >
+        <header class="modal-head">
+          <span class="modal-title">{fileName()}</span>
+          <span class="modal-path">{props.path}</span>
+          <button
+            class="modal-button"
+            onClick={() => props.onOpenInOs(props.path)}
+            title="Open in OS default viewer"
+          >
+            ↗
+          </button>
+          <button class="modal-button" onClick={props.onClose} title="Close (Esc)">
+            ×
+          </button>
+        </header>
+        <div class="pdf-body">
+          <webview src={fileUrl()} partition="persist:pdf" class="pdf-webview" />
+        </div>
+      </div>
+    </div>
+  );
+}
